# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Kasragay API
  description: |-
    ## Notes 
    ---
    - **09202400120** phone number is signed up by default as admin, and client.
      - **Admin ID** 89950e97-6b0f-4ef4-977a-8378b14bc4a7
      - **Client ID** 779033a2-4eaa-4817-aa81-24e24bd419f5
    - **09202400120-9** phone numbers are by-passed and available for testing.

    ### Logs
    ---
    **2025-07-09**
    - Initial commit
  version: 1.0.0
servers:
  - url: https://api.kasragay.com/v1
    description: API Server
...
paths:
  /auth/check:
    get:
      tags:
        - auth
      summary: Checks user is already registered or not (10 r/m)
      description: Checks Whether the provided identity is already registered or not
      parameters:
        - name: username
          in: query
          required: true
          description: Username to check
          schema:
            type: string
            example: "Kasra"
            pattern: ^(?!.*\.\.)[a-zA-Z0-9\-_]+(\.[a-zA-Z0-9\-_]+)*$
            minLength: 3
            maxLength: 64
        - name: user_type
          in: query
          required: true
          description: User Type to check
          schema:
            type: string
            example: "client"
            enum: [client, admin]
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCheckPostResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /auth/signup/key:
    get:
      tags:
        - auth
      summary: Send Key (3 r/m)
      description: Send an Key to user's phone number
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          description: Email of user
          required: false
          schema:
            type: string
            example: "amirhossein.pdn@gmail.com"
            format: email
        - name: phone_number
          in: query
          description: Phone number of user
          required: false
          schema:
            type: string
            example: "09202400120"
            pattern: ^0\d{10}$
        - name: user_type
          in: query
          description: User Type of user
          required: true
          schema:
            type: string
            example: "admin"
            enum: [admin]
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtUnauthorizedResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "409":
          description: Username already exists Or PhoneNumber Or Email is taken by multiple users
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UsernameAlreadyExistsResponse"
                  - $ref: "#/components/schemas/EmailIsTakenByMultipleResponse"
                  - $ref: "#/components/schemas/PhoneNumberIsTakenByMultipleResponse"
        "410":
          description: User has been deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDeletedResponse"
        "429":
          description: Too Early Or Too many requests
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AuthMethodOtpGetTooEarlyResponse"
                  - $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /auth/tmp/signup/key:
    get:
      tags:
        - auth
      summary: Get Send Key (3 r/m)
      description: Get send an Key to user's phone number
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          description: Email of user
          required: false
          schema:
            type: string
            example: "amirhossein.pdn@gmail.com"
            format: email
        - name: phone_number
          in: query
          description: Phone number of user
          required: false
          schema:
            type: string
            example: "09202400120"
            pattern: ^0\d{10}$
        - name: user_type
          in: query
          description: User Type of user
          required: true
          schema:
            type: string
            example: "admin"
            enum: [admin]
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmpAuthAdminSignupKeyGetResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtUnauthorizedResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /auth/{otpType}/otp:
    get:
      tags:
        - auth
      summary: Send OTP (3 r/m)
      description: Send an OTP to user's phone number
      parameters:
        - name: otpType
          in: path
          description: Otp type
          required: true
          schema:
            type: string
            example: "sign-up"
            enum:
              [
                sign-up,
                sign-in,
                change-password,
                change-phone,
                delete-account,
                change-recovery-email,
              ]
        - name: username
          in: query
          required: true
          description: Username to check
          schema:
            type: string
            example: "Kasra"
            pattern: ^(?!.*\.\.)[a-zA-Z0-9\-_]+(\.[a-zA-Z0-9\-_]+)*$
            minLength: 3
            maxLength: 64
        - name: send_to_email
          in: query
          description: Send to email of user
          required: false
          schema:
            type: boolean
            example: false
        - name: send_to_phone
          in: query
          description: Send to phone number of user
          required: false
          schema:
            type: boolean
            example: false
        - name: user_type
          in: query
          description: User Type of user
          required: true
          schema:
            type: string
            example: "client"
            enum: [client, admin]
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMethodOtpGetResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UserHasNotSetEmailResponse"
                  - $ref: "#/components/schemas/UserHasNotSetPhoneNumberResponse"
                  - $ref: "#/components/schemas/BadRequestResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameNotFoundResponse"
        "409":
          description: Username already exists Or PhoneNumber Or Email is taken by multiple users
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UsernameAlreadyExistsResponse"
                  - $ref: "#/components/schemas/EmailIsTakenByMultipleResponse"
                  - $ref: "#/components/schemas/PhoneNumberIsTakenByMultipleResponse"
        "410":
          description: User has been deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDeletedResponse"
        "429":
          description: Too Early Or Too many requests
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AuthMethodOtpGetTooEarlyResponse"
                  - $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /auth/tmp/{otpType}/otp:
    get:
      tags:
        - auth
      summary: Get Send OTP (3 r/m)
      description: Get send an OTP to user's phone number
      parameters:
        - name: otpType
          in: path
          description: Otp type
          required: true
          schema:
            type: string
            example: "sign-up"
            enum:
              [
                sign-up,
                sign-in,
                change-password,
                change-phone,
                delete-account,
                change-recovery-email,
              ]
        - name: username
          in: query
          required: true
          description: Username to check
          schema:
            type: string
            example: "Kasra"
            pattern: ^(?!.*\.\.)[a-zA-Z0-9\-_]+(\.[a-zA-Z0-9\-_]+)*$
            minLength: 3
            maxLength: 64
        - name: send_to_email
          in: query
          description: Send to email of user
          required: false
          schema:
            type: boolean
            example: false
        - name: send_to_phone
          in: query
          description: Send to phone number of user
          required: false
          schema:
            type: boolean
            example: false
        - name: user_type
          in: query
          description: User Type of user
          required: true
          schema:
            type: string
            example: "client"
            enum: [client, admin]
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmpAuthMethodOtpGetResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameNotFoundResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /auth/signup:
    post:
      tags:
        - auth
      summary: Signup using OTP (10 r/m)
      description: Signup using OTP which has been sent to user's phone number
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthSignupPostRequest"
      responses:
        "200":
          description: Successful Login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Login"
        "400":
          description: Bad request or Bad Avatar
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/BadRequestResponse"
                  - $ref: "#/components/schemas/BadAvatarResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenIncorrectResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "409":
          description: Username already exists Or PhoneNumber Or Email is taken by multiple users
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/UsernameAlreadyExistsResponse"
                  - $ref: "#/components/schemas/EmailIsTakenByMultipleResponse"
                  - $ref: "#/components/schemas/PhoneNumberIsTakenByMultipleResponse"
        "415":
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /auth/signin:
    post:
      tags:
        - auth
      summary: Signin using OTP (10 r/m)
      description: Signin using OTP
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthSigninPostRequest"
      responses:
        "200":
          description: Successful Login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Login"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenIncorrectResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameNotFoundResponse"
        "410":
          description: User has been deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDeletedResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthCheckPostResponse:
      type: object
      required:
        - exists
        - deleted
        - has_email
        - has_phone_number
        - has_password
      properties:
        exists:
          type: boolean
          example: true
        deleted:
          type: boolean
          example: false
        has_email:
          type: boolean
          example: false
        has_phone_number:
          type: boolean
          example: false
        has_password:
          type: boolean
          example: false
...
    InternalServerErrorResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1000
          enum: [1000]
        message:
          type: string
          example: "internal server error"
...
    InvalidHostResponse:
      type: object
      required:
        - status
        - code
        - message
        - reasons
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1002
          enum: [1002]
        message:
          type: string
          example: "inavlid host"
        reasons:
          type: object
          required:
            - host
          properties:
            host:
              type: string
              example: "88.101.15.11:443"
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...
    BadRequestResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1004
          enum: [1004]
        message:
          type: string
          example: "bad request"
        reasons:
          type: object
          properties:
            field1:
              type: string
              example: "value1"
            field2:
              type: string
              example: "value2"

        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...